Add react-router and home page
