+++ src/webapp/src/utils/ajax_get.js
@@ -1,0 +1,24 @@
+const API_ENDPOINT = 'http://localhost:8080';
+
+const ajax_get = function(url, callback) {
+  var xmlhttp = new XMLHttpRequest();
+  xmlhttp.onreadystatechange = function() {
+    if (xmlhttp.readyState == 4 && xmlhttp.status == 200) {
+      console.log('responseText:' + xmlhttp.responseText);
+      try {
+        var data = JSON.parse(xmlhttp.responseText);
+      } catch (err) {
+        console.log(err.message + ' in ' + xmlhttp.responseText);
+        return;
+      }
+      callback(data);
+    }
+  };
+
+  xmlhttp.open('GET', url, true);
+  xmlhttp.send();
+};
+
+export { API_ENDPOINT };
+
+export default ajax_get;
